Modify type of allocatedStorage for database configuration to be an integer rather than a double

+++ infrastructure/src/main/java/institute/hopesoftware/ApplicationStack.java
@@ -211,7 +211,7 @@ public class ApplicationStack extends Stack {
 
         String postgresVersion = dbConfiguration.getPostgresVersion();
         String dbInstanceClass = dbConfiguration.getDbInstanceClass();
-        double allocatedStorage = dbConfiguration.getAllocatedStorage();
+        Integer allocatedStorage = dbConfiguration.getAllocatedStorage();
 
         dbInstance = CfnDBInstance.Builder.create(this, "postgresInstance")
                 .dbInstanceIdentifier(applicationEnvironment.prefix("database"))
